Move SiteInfo to be shared component
